--- CarEshipApp/app/src/main/java/com/example/careshipapp/UpdatePassword.java
@@ -1,5 +1,0 @@
-package com.example.careshipapp;
-
-public class UpdatePassword {
-
-}
