+++ CarEshipApp/app/src/main/java/com/example/careshipapp/car_control/JoystickMainActivity.java
@@ -1,4 +1,4 @@
-package com.example.careshipapp.CarControl;
+package com.example.careshipapp.car_control;
 
 import android.os.Bundle;
