+++ CarEshipApp/app/src/main/java/com/example/careshipapp/customer_gui/CommonCold.java
@@ -1,4 +1,4 @@
-package com.example.careshipapp;
+package com.example.careshipapp.customer_gui;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.careshipapp.R;
 
 import java.util.ArrayList;
 import java.util.Collections;
