+++ CarEshipApp/app/src/main/java/com/example/careshipapp/customer_gui/CustomerAddress.java
@@ -1,13 +1,14 @@
-package com.example.careshipapp;
+package com.example.careshipapp.customer_gui;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.careshipapp.user_management.DBHelperClass;
+import com.example.careshipapp.R;
 
 
 public class CustomerAddress extends AppCompatActivity {
