+++ CarEshipApp/app/src/main/java/com/example/careshipapp/customer_gui/MainActivityCategoryList.java
@@ -1,8 +1,10 @@
-package com.example.careshipapp;
+package com.example.careshipapp.customer_gui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.careshipapp.R;
 
 public class MainActivityCategoryList extends AppCompatActivity {
