+++ CarEshipApp/app/src/main/java/com/example/careshipapp/user_management/MainActivityCreateAccount.java
@@ -1,4 +1,4 @@
-package com.example.careshipapp;
+package com.example.careshipapp.user_management;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -8,6 +8,7 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.careshipapp.R;
 import com.google.android.material.button.MaterialButton;
 
 public class MainActivityCreateAccount extends AppCompatActivity {
