+++ CarEshipApp/app/src/main/res/layout/activity_main_category_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@null"
-    tools:context=".MainActivityCategoryList">
+    tools:context=".customer_gui.MainActivityCategoryList">
 
 
     <ImageView
