+++ CarEshipApp/app/src/main/res/layout/activity_main_create_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@null"
-    tools:context=".MainActivityCreateAccount">
+    tools:context=".user_management.MainActivityCreateAccount">
 
     <ImageView
         android:id="@+id/imageView"
