+++ CarEshipApp/app/src/main/res/layout/allergies.xml
@@ -7,7 +7,7 @@
 
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".heartdisease">
+    tools:context=".customer_gui.heartdisease">
 
     <TextView
         android:id="@+id/medicine_order"
