+++ CarEshipApp/app/src/main/res/layout/anxiety.xml
@@ -7,7 +7,7 @@
 
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context="anxiety">
+    tools:context=".customer_gui.anxiety">
 
     <TextView
         android:id="@+id/medicine_order"
