+++ CarEshipApp/app/src/main/res/layout/asthma.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".asthma">
+    tools:context=".customer_gui.asthma">
 
     <TextView
         android:id="@+id/medicine_order"
