+++ CarEshipApp/app/src/main/res/layout/commoncold.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".CommonCold">
+    tools:context=".customer_gui.CommonCold">
 
     <TextView
         android:id="@+id/medicine_order"
