+++ CarEshipApp/app/src/main/res/layout/heartdisease.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".heartdisease">
+    tools:context=".customer_gui.heartdisease">
 
     <TextView
         android:id="@+id/medicine_order"
