+++ CarEshipApp/app/src/main/res/layout/staff_login_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@null"
-    tools:context=".MainActivityLoginPage">
+    tools:context=".user_management.MainActivityLoginPage">
 
     <ImageView
         android:id="@+id/imageView"
